test: refactor folders
